requirements.txt
pandas
plotly
opencv-python
av
ffmpeg-python
tqdm
moviepy
vidgear
google-cloud-videointelligence
pymediainfo
Pillow
PyYAML
torch
torchvision
torchaudio
ultralytics
tensorflow
transformers
einops
timm
tqdm
matplotlib
pipreqs
pylint
numpy
python_dateutil